AI Playground/requirements.txt
streamlit
google-cloud-generative-ai
PyPDF2
docx2txt